# backend/src/docs/swagger/swagger-main.yaml
openapi: 3.0.0
info:
  title: Restaurant Management API - Complete
  description: |
    Complete API documentation for the Restaurant Management System.

    **Available Modules:**
    - 🔐 Authentication & User Management (3 endpoints)
    - 🔑 Password Management (3 endpoints)
    - 🍔 Menu Management (8 endpoints)
    - 📦 Order Management (6 endpoints)
    - 📊 Dashboard Analytics (1 endpoint)
    - 💬 Facebook Integration (5 endpoints)
    - 📱 WhatsApp Integration (6 endpoints)

    **Total: 32 Endpoints**
  version: 2.0.0
  contact:
    name: API Support
    email: support@restaurant-api.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://event-bot-admin.chotkari.com/api
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token. Format - Bearer {token}
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Error message"

    # ==================== Auth Schemas ====================
...
    RegisterStaffRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          example: "jane_staff"
        email:
          type: string
          example: "jane@example.com"
        password:
          type: string
          example: "StaffPass123!"

          # User Schemas
...
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /auth/register-staff:
    post:
      tags: [Authentication]
      summary: Register staff member (Owner only)
      description: Register staff for the restaurant. Owner only.
      operationId: registerStaff
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterStaffRequest"
      responses:
        "201":
          description: Staff registered
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
...
  /restaurant/staff/{id}:
    delete:
      tags: [Users]
      summary: Delete staff member
      description: Permanently delete a staff member (hard delete)
      operationId: deleteStaff
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Staff member ID
          schema:
            type: integer
            example: 5
      responses:
        "200":
          description: Staff member deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Staff member John Doe has been deleted successfully
        "400":
          description: Cannot delete self
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /restaurant/staff/{id}/deactivate:
    patch:
      tags: [Users]
      summary: Deactivate staff member
      description: Soft delete - deactivate staff member without removing data
      operationId: deactivateStaff
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Staff member ID
          schema:
            type: integer
            example: 5
      responses:
        "200":
          description: Staff member deactivated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          description: Cannot deactivate self
        "404":
          $ref: "#/components/responses/NotFoundError"
  /restaurant/staff/{id}/reactivate:
    patch:
      tags: [Users]
      summary: Reactivate staff member
      description: Restore a previously deactivated staff member
      operationId: reactivateStaff
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Staff member ID
          schema:
            type: integer
      responses:
        "200":
          description: Staff member reactivated successfully
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ==================== Password Management Endpoints ====================
...
security: []